+++ lib/features/foodboxes/presentation/screen/boxes_screen.dart
@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:get_it/get_it.dart';
 import 'package:sliver_tools/sliver_tools.dart';
 import 'package:zachranobed/common/constants.dart';
+import 'package:zachranobed/common/helper_service.dart';
 import 'package:zachranobed/common/utils/date_time_utils.dart';
 import 'package:zachranobed/extensions/build_context_extensions.dart';
+import 'package:zachranobed/features/foodboxes/domain/repository/food_box_repository.dart';
 import 'package:zachranobed/features/foodboxes/presentation/widget/box_movement_list.dart';
 import 'package:zachranobed/ui/widgets/button.dart';
+import 'package:zachranobed/ui/widgets/empty_page.dart';
+import 'package:zachranobed/ui/widgets/error_page.dart';
+import 'package:zachranobed/ui/widgets/loading_page.dart';
 
 class BoxesScreen extends StatefulWidget {
   const BoxesScreen({super.key});
@@ -15,13 +21,52 @@ class BoxesScreen extends StatefulWidget {
 
 class _BoxesScreenState extends State<BoxesScreen> {
   final List<DateTime> _previousWeeks = [];
+  late Future<int> _boxMovementCountFuture;
 
   final DateTime _thisWeekStart = DateTimeUtils.getWeekStart(DateTime.now());
   late final DateTime _previousWeekStart =
       DateTimeUtils.getPreviousWeek(_thisWeekStart);
 
   @override
+  void initState() {
+    super.initState();
+    final repository = GetIt.I<FoodBoxRepository>();
+    _boxMovementCountFuture = repository.getMovementBoxesCount(
+      entityId: HelperService.getCurrentUser(context)!.entityId,
+    );
+  }
+
+  @override
   Widget build(BuildContext context) {
+    final pageTitle = context.l10n!.boxes;
+    return FutureBuilder<int>(
+      future: _boxMovementCountFuture,
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return LoadingPage(pageTitle: pageTitle);
+        } else if (snapshot.hasError) {
+          return ErrorPage(
+            pageTitle: pageTitle,
+            error: snapshot.error,
+          );
+        } else {
+          final boxCount = snapshot.data!;
+          return boxCount < 1
+              ? Scaffold(
+                  appBar: AppBar(title: Text(pageTitle)),
+                  body: EmptyPage(
+                    vectorImagePath: ZOStrings.boxEmptyPath,
+                    title: context.l10n!.boxesEmptyTitle,
+                    description: context.l10n!.boxesEmptyDescription,
+                  ),
+                )
+              : _buildContent(context);
+        }
+      },
+    );
+  }
+
+  Widget _buildContent(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text(context.l10n!.boxes),