+++ lib/features/offeredfood/presentation/screens/donations_screen.dart
@@ -1,10 +1,16 @@
 import 'package:flutter/material.dart';
+import 'package:get_it/get_it.dart';
 import 'package:sliver_tools/sliver_tools.dart';
 import 'package:zachranobed/common/constants.dart';
+import 'package:zachranobed/common/helper_service.dart';
 import 'package:zachranobed/common/utils/date_time_utils.dart';
 import 'package:zachranobed/extensions/build_context_extensions.dart';
-import 'package:zachranobed/ui/widgets/button.dart';
+import 'package:zachranobed/features/offeredfood/domain/repository/offered_food_repository.dart';
 import 'package:zachranobed/features/offeredfood/presentation/widget/donated_food_list.dart';
+import 'package:zachranobed/ui/widgets/button.dart';
+import 'package:zachranobed/ui/widgets/empty_page.dart';
+import 'package:zachranobed/ui/widgets/error_page.dart';
+import 'package:zachranobed/ui/widgets/loading_page.dart';
 
 class DonationsScreen extends StatefulWidget {
   const DonationsScreen({super.key});
@@ -15,23 +21,60 @@ class DonationsScreen extends StatefulWidget {
 
 class _DonationsScreenState extends State<DonationsScreen> {
   final List<DateTime> _previousWeeks = [];
+  late Future<int> _mealsCountFuture;
 
   final DateTime _thisWeekStart = DateTimeUtils.getWeekStart(DateTime.now());
   late final DateTime _previousWeekStart =
       DateTimeUtils.getPreviousWeek(_thisWeekStart);
 
   @override
+  void initState() {
+    super.initState();
+    final repository = GetIt.I<OfferedFoodRepository>();
+    _mealsCountFuture = repository.getSavedMealsCount(
+      entityId: HelperService.getCurrentUser(context)!.entityId,
+      timePeriod: null,
+    );
+  }
+
+  @override
   Widget build(BuildContext context) {
+    final pageTitle = context.l10n!.food;
+    return FutureBuilder<int>(
+      future: _mealsCountFuture,
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          return LoadingPage(pageTitle: pageTitle);
+        } else if (snapshot.hasError) {
+          return ErrorPage(
+            pageTitle: pageTitle,
+            error: snapshot.error,
+          );
+        } else {
+          final mealsCount = snapshot.data!;
+          return mealsCount < 1
+              ? Scaffold(
+                  appBar: AppBar(title: Text(pageTitle)),
+                  body: EmptyPage(
+                    vectorImagePath: ZOStrings.chefEmptyPath,
+                    title: context.l10n!.donationsEmptyTitle,
+                    description: context.l10n!.donationsEmptyDescription,
+                  ),
+                )
+              : _buildContent(context);
+        }
+      },
+    );
+  }
+
+  Widget _buildContent(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(context.l10n!.food),
-      ),
+      appBar: AppBar(title: Text(context.l10n!.food)),
       body: CustomScrollView(
         slivers: [
           SliverPadding(
-            padding: const EdgeInsets.symmetric(
-              horizontal: WidgetStyle.padding,
-            ),
+            padding:
+                const EdgeInsets.symmetric(horizontal: WidgetStyle.padding),
             sliver: MultiSliver(
               children: [
                 DonatedFoodList(